fix airlift bug
A player in the posession of an airlift card should be able to airlift
themselves without having to create a movement proposal and they should
be able to do that at any time. They should be able to create a movement
proposal at any time when trying to move another player.